test(freshness_checksum): verify cargo depinfo is properly encoded
